// src/app/model/playlist.ts
export class PlaylistRequest {
    id: number = 0;
    postedby: number = 0;
    title: string = "";
    accountid: number = 0;
    category: string = "";
    nameofsong: string = "";
    description: string = "";
    songid: number = 0;
    favourite: any;


}
export class PlaylistResponse {

    postedby: number = 0;
    title: string = "";
    category: string = "";
    nameofsong: string = "";
    description: string = "";
    songid: number = 0;
}